unit-test: move test for filtered queries to another class
this is to make tests run synchronously and not mix up the queries. allowing those to coexist in the same class allows them to run concurrently, thus there are data races in database queries

the last statement is NOT a problem, as the only problem with those data races is cleaning up database while the other coroutine expects it to be populated

throwing tests in different classes allows us to execute them in different loops, which are spawned one after the other. while slow, it allows to run testcases after other finish. this is testing code and will not affect the end product, so it is fine to make it suboptimal