unit-tests: minor modifications to unit tests
even though it seems a lot to be called "minor", most of those is just additions of sorts and whatnot. those are needed as database relies extra heavily on asynchronimity to keep the project from ultimate demise due to horribly slow code. there is however a removed test, but its functionality is implemented in a better one that\'s not so much controlled however